refactor(app): move solver-shared helpers into solvers/
Moves websocket.ts, line-plot.tsx and range.ts out of the flat shared/
folder into solvers/ (next to the existing solver-agnostic
boltzmann.ts), since they are generic infrastructure meant to be
reused across solver implementations, not shared/ grab-bag utilities.
Currently bolsig is the only solver consuming them. Updates import
paths accordingly. No behavioral changes.
