+++ app/src/solvers/line-plot.tsx
